Use uint64_t for binary prefixes since size_t is 32 bit in Wasm

+++ core/silkworm/common/util.cpp
@@ -143,7 +143,7 @@ Bytes from_hex(std::string_view hex) {
     return out;
 }
 
-std::optional<size_t> parse_size(const std::string& sizestr) {
+std::optional<uint64_t> parse_size(const std::string& sizestr) {
     if (sizestr.empty()) {
         return 0;
     }
